move interop tests back
