+++ test/interop/interop_client.js
@@ -20,10 +20,9 @@
 
 var fs = require('fs');
 var path = require('path');
-// TODO(murgatroid99): use multiple grpc implementations
-var grpc = require('..');
+var grpc = require('../../packages/grpc-native-core');
 var testProto = grpc.load({
-  root: __dirname + '/../deps/grpc',
+  root: __dirname + '/../../packages/grpc-native-core/deps/grpc',
   file: 'src/proto/grpc/testing/test.proto'}).grpc.testing;
 var GoogleAuth = require('google-auth-library');
 
@@ -566,7 +565,7 @@ function runTest(address, host_override, test_case, tls, test_ca, done, extra) {
   if (tls) {
     var ca_path;
     if (test_ca) {
-      ca_path = path.join(__dirname, '../test/data/ca.pem');
+      ca_path = path.join(__dirname, '../data/ca.pem');
       var ca_data = fs.readFileSync(ca_path);
       creds = grpc.credentials.createSsl(ca_data);
     } else {
